+++ src/mako/error/handlers/web/views/400.tpl.php
@@ -1,8 +1,8 @@
 {% extends:'mako-error::error' %}
 
-{% block:title %}Bad Request{% endblock %}
+{% block:'title' %}Bad Request{% endblock %}
 
-{% block:message %}
+{% block:'message' %}
 	<h1>400</h1>
 	<p>The server was unable to process the request.</p>
 {% endblock %}