+++ src/mako/error/handlers/web/views/403.tpl.php
@@ -1,8 +1,8 @@
 {% extends:'mako-error::error' %}
 
-{% block:title %}Forbidden{% endblock %}
+{% block:'title' %}Forbidden{% endblock %}
 
-{% block:message %}
+{% block:'message' %}
 	<h1>403</h1>
 	<p>You don't have permission to access the requested resource.</p>
 {% endblock %}