+++ src/mako/error/handlers/web/views/404.tpl.php
@@ -1,8 +1,8 @@
 {% extends:'mako-error::error' %}
 
-{% block:title %}Not Found{% endblock %}
+{% block:'title' %}Not Found{% endblock %}
 
-{% block:message %}
+{% block:'message' %}
 	<h1>404</h1>
 	<p>The resource you requested could not be found. It may have been moved or deleted.</p>
 {% endblock %}