+++ src/mako/error/handlers/web/views/410.tpl.php
@@ -1,8 +1,8 @@
 {% extends:'mako-error::error' %}
 
-{% block:title %}Gone{% endblock %}
+{% block:'title' %}Gone{% endblock %}
 
-{% block:message %}
+{% block:'message' %}
 	<h1>410</h1>
 	<p>The resource you requested is no longer available and will not be available again.</p>
 {% endblock %}