+++ src/mako/error/handlers/web/views/429.tpl.php
@@ -1,8 +1,8 @@
 {% extends:'mako-error::error' %}
 
-{% block:title %}Too Many Requests{% endblock %}
+{% block:'title' %}Too Many Requests{% endblock %}
 
-{% block:message %}
+{% block:'message' %}
 	<h1>429</h1>
 	<p>You have made too many requests to the server.</p>
 {% endblock %}