+++ src/mako/error/handlers/web/views/503.tpl.php
@@ -1,8 +1,8 @@
 {% extends:'mako-error::error' %}
 
-{% block:title %}Service Unavailable{% endblock %}
+{% block:'title' %}Service Unavailable{% endblock %}
 
-{% block:message %}
+{% block:'message' %}
 	<h1>503</h1>
 	<p>The service is currently unavailable.</p>
 {% endblock %}